revert api change, but add deprecate
since almost all other public api changes were reverted, I am
reverting this one as well since it isn't unreasonable to guess
that merely adding 'serializable' may not be significant enough of
an api change to require a larger release.

based on the usages and documentation of the constructor though, it
seemed like a logic change, so I added a Deprecation warning to
reflect our intent to change it.