fix random, erroneous hllp test failure
this was introduced a couple commits ago, but does not always occur.
It is a problem somewhat caused by the equals method being able to cause
changes in internal state that affected how we would have otherwise
expecting the equals method to behave. Since this internal state is not
visible to non-test code, and such changes aren't unheard of (although
admittedly more common for hashCode), we have modified the test case
to account for this possibility.